Add Sixel color, window, and cell dimension query support
This branch builds on and compares to #310 (as "base branch") which first introduced "does_sixel" method that is adjusted here (moved in docs and tests)

Sixel
-----

*New* files docs/sixel.rst and tests/test_sixel.py describe and exercise existing does_sixel() method and the following new public api methods,
  - get_sixel_height_and_width() - Gets maximum pixel dimensions for sixel graphics
  - get_cell_height_and_width() - Gets pixel dimensions of a single character cell
  - get_sixel_colors() - Determines available color registers for sixel graphics

Inquiry Guardrail
-------------------

*New* common "guardrail",  moved the sometimes commonly used ``does_styling and is_a_tty`` code check to base private method _query_response(), avoiding long or sometimes indefinite timeouts with non-terminals with the following methods:
  - get_location()
  - get_fgcolor()
  - get_bgcolor()

Further, do not check ``does_styling`` anymore in this common guardrail when making inquiries to the Terminal -- yes, we may not wish to do colors and sequences, but, this shouldn't stop us from enabling keyboard kitty mode, bracketed paste, or inquiring about window dimensions.

Tests have been updated to match, requiring special 'is_a_tty=True' argument to pass for these tests -- we should not have "inquired about a terminal" with these sequences unless is_a_tty is True, so this is something of a change or bugfix.