Add inband-resize decoding and helpers
This branch builds on and compares to #312 (as "base branch") which is required for some common code.

Suggest a new way to detect resize, according to this Specification, https://gist.github.com/rockorager/e695fb2924d36b2bcf1fff4a3704bd83 detected by DEC Private mode 2048, with 'EVENT_RESIZE' keystroke name returned by inkey.

This is a bit new and not popularly supported, but it is a good solution to the protocol and system-independent support for window resize events. For example, this would work over serial where no existing protocols would.